+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Unreleased
 
-TK
+### New features
+
+- Support updates of settings plugins in `settings.gradle.kts` and `settings.gradle` files (including refreshVersions itself).
 
 ## Version 0.9.7 (2020-10-16)